Added 'edit' function and updated README file
