+++ src/index.html
@@ -1,13 +1,14 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>To Do List</title>
 </head>
-<body>
 
+<body>
     <section>
         <h2 id="title">Today's To Do</h2>
         <input type="text" id="input-field" placeholder="Add to your list..." required>
@@ -16,6 +17,6 @@
         </div>
         <h2 id="clear-button">Clear all completed</h2>
     </section>
-
 </body>
+
 </html>
